Improve server side socket event handling
